requirements.txt
flask==3.0.0
flask-jwt-extended==4.5.3
pymongo==4.5.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.24.3
deepface==0.0.79
streamlit==1.28.0
Pillow==10.0.0
requests==2.31.0
scikit-learn==1.3.2
